Make FLCE compatible with FSDP and PEFT (#674)
## Summary

FLCE currently fails to get the lm_head module weights when PEFT and/or
FSDP is used.

This happens when the lm_head module is trainable, e.g., when it's added
to the `modules_to_save` list for parameter-efficient fine-tuning with
LoRA, or when full fine-tuning is distributed with FSDP.

This PR makes sure the lm_head module weights can be read in all
configurations, by unwrapping PEFT `ModulesToSaveWrapper` and FSDP
`FullyShardedDataParallel` wrappers, and calling the later within FDSP
forward pass.

Fixes #48
Fixes #332 

## Testing Done

Successfully tested SFT with PEFT/LoRA and FSDP with the following
configuration:

```
model_name_or_path: Meta-Llama/Meta-Llama-3.1-8B-Instruct
model_revision: main
torch_dtype: bfloat16
attn_implementation: flash_attention_2
use_liger: [false, true]

use_peft: true
lora_r: 16
lora_alpha: 8
lora_dropout: 0.05
lora_target_modules: ["q_proj", "v_proj", "k_proj", "o_proj", "gate_proj", "up_proj", "down_proj"]
lora_modules_to_save: ["lm_head", "embed_tokens"]

fsdp: "full_shard auto_wrap"
fsdp_config:
  activation_checkpointing: true
  cpu_ram_efficient_loading: false
  sync_module_states: true
  use_orig_params: true
  limit_all_gathers: false
```

The train loss curves with and without Liger:


![image](https://github.com/user-attachments/assets/63a1a26c-70af-4ed4-a12b-425ee04df69e)

I've also successfully tested full fine-tuning with FSDP, i.e., the same
configuration as above except but `use_peft: false`, which gives the
following train loss curves with and without Liger:


![image](https://github.com/user-attachments/assets/9811d893-8ec5-4589-b8e9-781fd7074257)

- Hardware Type: 8x NVIDIA-A100-SXM4-80GB
- [ ] run `make test` to ensure correctness
- [x] run `make checkstyle` to ensure code style
- [ ] run `make test-convergence` to ensure convergence
